# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Book API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/mentions": {
            "get": {
                "description": "\"Mentioned Moments\" — every Moment the authenticated user has ever been mentioned in",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mentions"
                ],
                "summary": "List moments I'm mentioned in",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Moments per page (default 20, max 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_ListMomentsResponse"
                        }
                    }
                }
            }
        },
...
        "/moments/{id}/mentions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mentions"
                ],
                "summary": "List a moment's mentions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Moment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_ListMentionsResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    }
                }
            },
            "post": {
                "description": "Names another user present in a personal Moment; automatically shares the Moment with them",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mentions"
                ],
                "summary": "Mention a user on a moment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Moment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Username to mention",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.CreateMentionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_MentionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    }
                }
            }
        },
        "/moments/{id}/mentions/leave": {
            "post": {
                "description": "The mentioned user removing themselves; the Moment stays with its owner, unchanged, and no notification is sent",
                "tags": [
                    "mentions"
                ],
                "summary": "Leave a mention (mentioned user)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Moment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    }
                }
            }
        },
        "/moments/{id}/mentions/{mentionId}": {
            "delete": {
                "description": "The Moment owner removing a mention they added by mistake",
                "tags": [
                    "mentions"
                ],
                "summary": "Delete a mention (owner)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Moment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Mention ID",
                        "name": "mentionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    }
                }
            }
        },
...
        "/moments/{id}/share": {
            "post": {
                "description": "The mention flow's optional \"Share to circle too?\" step — never a standalone share action",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mentions"
                ],
                "summary": "Share a moment to a circle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Moment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Target circle",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.ShareMomentToCircleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_MomentCircleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    }
                }
            }
        },
        "/moments/{id}/share/{circleId}": {
            "delete": {
                "tags": [
                    "mentions"
                ],
                "summary": "Unshare a moment from a circle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Moment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Circle ID",
                        "name": "circleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.CreateMentionRequest": {
            "type": "object",
            "required": [
                "username"
            ],
            "properties": {
                "username": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 3
                }
            }
        },
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.ListMentionsResponse": {
            "type": "object",
            "properties": {
                "mentions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.MentionResponse"
                    }
                }
            }
        },
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.ListMomentsResponse": {
            "type": "object",
            "properties": {
                "moments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.MomentResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.PaginationResponse"
                }
            }
        },
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.MentionResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-07-20T10:00:00Z"
                },
                "display_name": {
                    "type": "string",
                    "example": "Gede"
                },
                "id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "mentioned_user_id": {
                    "type": "string"
                },
                "moment_id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                }
            }
        },
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.MomentCircleResponse": {
            "type": "object",
            "properties": {
                "circle_id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "moment_id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "shared_at": {
                    "type": "string",
                    "example": "2026-07-20T10:00:00Z"
                },
                "shared_by_user_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.MomentResponse": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "color_hex": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-08-04T14:35:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "note": {
                    "type": "string"
                },
                "occurred_at": {
                    "description": "OccurredAt is reconstructed from occurred_local +\noccurred_utc_offset_minutes, so it round-trips exactly the wall\ntime and offset the Moment was captured with.",
                    "type": "string",
                    "example": "2026-08-04T14:30:00+09:00"
                },
                "occurred_utc_offset_minutes": {
                    "type": "integer",
                    "example": 540
                },
                "origin": {
                    "type": "string",
                    "example": "manual"
                },
                "place_name": {
                    "type": "string"
                },
                "recorded_at": {
                    "type": "string",
                    "example": "2026-08-04T14:35:00Z"
                },
                "settling_time_seconds": {
                    "type": "integer",
                    "example": 300
                },
                "thread_id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-08-04T14:35:00Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                }
            }
        },
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.PaginationResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 20
                },
                "total": {
                    "type": "integer",
                    "example": 42
                }
            }
        },
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.ShareMomentToCircleRequest": {
            "type": "object",
            "required": [
                "circle_id"
            ],
            "properties": {
                "circle_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_ListMentionsResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.ListMentionsResponse"
                },
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_ListMomentsResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.ListMomentsResponse"
                },
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_MentionResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.MentionResponse"
                },
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_MomentCircleResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.MomentCircleResponse"
                },
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
...
    }
}